Use cob-app/get to back tictactoe-server
This means that we don't have to re-implement functionality to get
the stylesheets and javascript files to load. While refactoring, I
fixed a quick display bug in core.js and moved the response format
over to the collection-based on from cob-app. I also deleted the
unused branch of the uril/respond function, added style.css for
testing purposes, and added an initializer in the responder.